pyproject.toml
[project]
name = "drf-spa-habits"
version = "0.1.0"
description = ""
authors = [
    {name = "Egor Farkov",email = "fef84@icloud.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
mypy = "^1.17.1"
flake8 = "^7.3.0"
isort = "^6.0.1"
black = "^25.1.0"

[tool.poetry.dependencies]
python = "^3.11"
